Keep Finder directory checks outside the sidebar row
The row now stores only a small cache while path expansion and off-main existence checks live in a dedicated helper. This keeps the deleted-directory behavior intact without growing ContentView beyond the CI file-length budget.

Constraint: ContentView.swift has a strict file-length budget enforced by CI
Rejected: Leave the async cache helper inline in ContentView | exceeded workflow guard budget by 27 lines
Confidence: medium
Scope-risk: narrow
Directive: Keep future workspace Finder path helpers outside ContentView unless the budget is reduced elsewhere
Tested: git diff --check; jq empty Resources/Localizable.xcstrings; plutil -lint GhosttyTabs.xcodeproj/project.pbxproj; wc -l Sources/ContentView.swift
Not-tested: local Swift tests or app launch before CI, per task instructions
